kexec/ifdown.c: Use AF_NETLINK instead of AF_INET
On embedded systems, there is no guarantee that INET support has been
built in.  However, as long as network support is built into the kernel,
we can use the netlink protocol to create a socket to still then use for
the ioctl calls since they don't require any certain family.  This will
help prevent EAFNOSUPPORT from being returned and printed to the console
on these systems.

Signed-off-by: Mason Rocha <mrocha@turretllc.us>
Signed-off-by: Simon Horman <horms@kernel.org>